Update Homepage, Overview, Design docs
