+++ docs/development/publishing.md
@@ -1,18 +1,17 @@
 ---
 nav: Publishing
-description: Share, publish, and sell your Packs.
+description: Share and publish your Packs.
 ---
 
 # Publishing your Pack
 
-Once your Pack is ready you can share it with others, publish it so it can be discovered, and even sell it. The available options depend on where your Pack will be used.
+Once your Pack is ready you can share it with others and publish it so it can be discovered. The available options depend on where your Pack will be used.
 
 ## Superhuman Docs Gallery
 
 Packs that work with :superhuman-docs: Docs should be published to the [Gallery][gallery]. They can be discovered there, as well as within the product itself. See the following articles for more information:
 
 - [Sharing & publishing][sharing_publishing]
-- [Selling Packs][selling_packs]
 
 
 ## Superhuman Go Store
@@ -24,6 +23,5 @@ Connectors that work with :superhuman-go: Go should be published to the [Store][
 
 [gallery]: https://coda.io/gallery?filter=packs
 [sharing_publishing]: https://help.coda.io/hc/en-us/articles/39555880940557-Share-or-publish-a-Pack
-[selling_packs]: https://help.coda.io/hc/en-us/articles/39555971447053-Sell-your-Packs-on-Coda
 [store]: https://superhuman.com/store/connectors
 [go_publishing]: https://docs.superhuman.com/@hector/go-agent-development-guide/listing-publishing-7
